Pull methods up into ApplicantData
