+++ server/app/services/CfJsonDocumentContext.java
@@ -26,7 +26,6 @@ import services.applicant.exception.JsonPathTypeMismatchException;
 import services.applicant.predicate.JsonPathPredicate;
 import services.applicant.question.Scalar;
 import services.export.JsonPrettifier;
-import services.geo.ServiceAreaInclusion;
 
 // NON_ABSTRACT_CLASS_ALLOWS_SUBCLASSING CfJsonDocumentContext
 
@@ -220,31 +219,6 @@ public class CfJsonDocumentContext {
   }
 
   /**
-   * Puts an array at a given path, building parent objects as needed.
-   *
-   * @param path the {@link Path} where the array should be added.
-   * @param entityNames a {@link List} containing service area results.
-   */
-  public void putServiceAreaInclusionEntities(
-      Path path, ImmutableList<ServiceAreaInclusion> entityNames) {
-    if (entityNames.isEmpty()) {
-      put(path, ImmutableList.of());
-    } else {
-      for (int i = 0; i < entityNames.size(); i++) {
-        putString(
-            path.atIndex(i).join(Scalar.SERVICE_AREA_ID.toDisplayString()),
-            entityNames.get(i).getServiceAreaId());
-        putString(
-            path.atIndex(i).join(Scalar.SERVICE_AREA_STATE.toDisplayString()),
-            entityNames.get(i).getState().name());
-        putLong(
-            path.atIndex(i).join(Scalar.SERVICE_AREA_TIMESTAMP.toDisplayString()),
-            entityNames.get(i).getTimeStamp());
-      }
-    }
-  }
-
-  /**
    * Puts null at the provided path.
    *
    * @param path a path
@@ -267,7 +241,7 @@ public class CfJsonDocumentContext {
    *     "$.applicant.children[3].favorite_color.text".
    * @param value the value to place; values of type Map will create the equivalent JSON structure
    */
-  private void put(Path path, Object value) {
+  protected void put(Path path, Object value) {
     checkLocked();
     putParentIfMissing(path);
     if (path.isArrayElement()) {
@@ -462,20 +436,6 @@ public class CfJsonDocumentContext {
     return this.readList(path, new TypeRef<ImmutableList<String>>() {});
   }
 
-  //
-  /**
-   * Attempt to read a list at the given {@link Path}. Returns {@code Optional#empty} if the path
-   * does not exist or a value other than an {@link ImmutableList} of {@link ServiceAreaInclusion}
-   * is found.
-   *
-   * @param path the {@link Path} to the list
-   * @return an Optional containing an ImmutableList<ServiceAreaInclusion>
-   */
-  public Optional<ImmutableList<ServiceAreaInclusion>> readServiceAreaList(Path path) {
-    return this.readList(
-        path.safeWithoutArrayReference(), new TypeRef<ImmutableList<ServiceAreaInclusion>>() {});
-  }
-
   /**
    * Attempt to read a list at the given {@link Path}. Returns {@code Optional#empty} if the path
    * does not exist or a value other than an {@link ImmutableList} of T is found.
@@ -485,7 +445,7 @@ public class CfJsonDocumentContext {
    * @param typeRef the {@link TypeRef} of the expected ImmutableList<T> type.
    * @return an Optional containing an ImmutableList<T>.
    */
-  private <T> Optional<ImmutableList<T>> readList(Path path, TypeRef<ImmutableList<T>> typeRef) {
+  protected <T> Optional<ImmutableList<T>> readList(Path path, TypeRef<ImmutableList<T>> typeRef) {
     try {
       return this.read(path, typeRef);
     } catch (JsonPathTypeMismatchException e) {
